full-stack mvc done

+++ src/main/java/edu/uncg/csc340/animal_api/PandaRepository.java
@@ -1,10 +1,13 @@
 package edu.uncg.csc340.animal_api;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
 
 import java.util.List;
 
-public interface PandaRepository extends JpaRepository<Panda, Integer> {
+public interface PandaRepository extends JpaRepository<Panda, Long> {
     List<Panda> findByType(String type);
     List<Panda> findByNameContaining(String name);
 }
+
